WIP: keyless leaf infrastructure (no_key_tag + partial specialization)
D2+D3 work in progress. Infrastructure compiles for non-keyless paths.
9 remaining errors in collision/split paths that call leaf->get_key_view()
on keyless leaves — these need key reconstruction from the inode path.

- art_common.hpp: no_key_tag, can_eliminate_key_in_leaf_v, leaf_key_type
- art_internal_impl.hpp: basic_leaf<no_key_tag, Header> specialization
- Deleter, unique_ptr, make_db_leaf_ptr updated for keyless leaves
- Scan/iterator cmp() guarded with if constexpr (category 2 done)
- OLC next()/prior() use keybuf_ for keyless leaves

NOT YET DONE: collision path key reconstruction (category 1)

+++ benchmark/micro_benchmark_key_view.cpp
@@ -85,7 +85,7 @@ void chain_insert(benchmark::State& state, key_view_set (*gen)(std::size_t)) {
     state.PauseTiming();
     unodb::benchmark::destroy_tree(db, state);
   }
-  state.SetItemsProcessed(static_cast<std::int64_t>(state.iterations()) *
+  state.SetItemsProcessed(state.iterations() *
                           static_cast<std::int64_t>(n));
 }
 
@@ -98,7 +98,7 @@ void chain_get(benchmark::State& state, key_view_set (*gen)(std::size_t)) {
   for (auto _ : state) {
     for (std::size_t i = 0; i < n; ++i) benchmark::DoNotOptimize(db.get(ks[i]));
   }
-  state.SetItemsProcessed(static_cast<std::int64_t>(state.iterations()) *
+  state.SetItemsProcessed(state.iterations() *
                           static_cast<std::int64_t>(n));
 }
 
@@ -115,7 +115,7 @@ void chain_remove(benchmark::State& state, key_view_set (*gen)(std::size_t)) {
     for (std::size_t i = 0; i < n; ++i)
       benchmark::DoNotOptimize(db.remove(ks[i]));
   }
-  state.SetItemsProcessed(static_cast<std::int64_t>(state.iterations()) *
+  state.SetItemsProcessed(state.iterations() *
                           static_cast<std::int64_t>(n));
 }
 
@@ -133,7 +133,7 @@ void chain_scan(benchmark::State& state, key_view_set (*gen)(std::size_t)) {
     });
     benchmark::DoNotOptimize(count);
   }
-  state.SetItemsProcessed(static_cast<std::int64_t>(state.iterations()) *
+  state.SetItemsProcessed(state.iterations() *
                           static_cast<std::int64_t>(n));
 }
 
@@ -155,7 +155,7 @@ void kv_vs_u64_insert(benchmark::State& state) {
     state.PauseTiming();
     unodb::benchmark::destroy_tree(db, state);
   }
-  state.SetItemsProcessed(static_cast<std::int64_t>(state.iterations()) *
+  state.SetItemsProcessed(state.iterations() *
                           static_cast<std::int64_t>(n));
 }
 
@@ -168,7 +168,7 @@ void kv_vs_u64_get(benchmark::State& state) {
   for (auto _ : state) {
     for (std::size_t i = 0; i < n; ++i) benchmark::DoNotOptimize(db.get(ks[i]));
   }
-  state.SetItemsProcessed(static_cast<std::int64_t>(state.iterations()) *
+  state.SetItemsProcessed(state.iterations() *
                           static_cast<std::int64_t>(n));
 }
 
@@ -185,7 +185,7 @@ void kv_vs_u64_remove(benchmark::State& state) {
     for (std::size_t i = 0; i < n; ++i)
       benchmark::DoNotOptimize(db.remove(ks[i]));
   }
-  state.SetItemsProcessed(static_cast<std::int64_t>(state.iterations()) *
+  state.SetItemsProcessed(state.iterations() *
                           static_cast<std::int64_t>(n));
 }
